Arc of Gauge to expand on mouseover
I noticed that the mouseover effect of expanding and contracting was not happening with the arc gauge. I thought it was working before, and I don't know if it was taken out on purpose of left out during the refactoring of the arc portion of the code. Here is a fix that goes with the current code.