Fix for #309
The original behavior prior to 2deefb3 had no interval wait times set to process the transitions. When bypassing the wait, the draw of the graph no longer has intermittent failure due to the intervals across graphs interfering with each other. It feels backwards for the default behavior to be the conditional, but I think it would be more code to reverse it at this time.