Explicitly set auto width for tooltip
c3's tooltip depends on the default behavior for table table width. However,
site stylesheets and CSS frameworks will often change the default table width.
This can completely break the rendering of the tooltip (e.g. when table width is
set to 100%, the tooltip will stretch from a left position in the negative
thousands all the way to the right of the chart). Explicitly setting
`width:auto;` for the tooltip allows c3 to play nicely with CSS frameworks and
site stylesheets.