Set the X axis in subchart to visible or invisible
I'm totally new in the github and all this stuff. Sorry if I do something wrong. 

I came across c3 for two days and found it suitable for me. Thanks for that!

But there are some points which I would like to improve. Here I have added two lines to the c3.js
and with following options the X axis could be set to hide
```
...
axis: {
  x: {
    show: false
  }
}
...
```
http://jsfiddle.net/vladsu/89szjf3r/3/ 

Best wishes,
Vlad