Keep the original index after a xSort
Keep the original index after a xSort, and add it to be the fifth parameter of tooltip format values:
tooltip: {
  format: {
    value: function(value,ratio,id,index,ori_index){
      return value;
    }
  }
}

Use case: Let's say X-axis is made of multiple variables. Ex. 'a+b'. We want to break down the data point on hover to show value of 'a' and 'b' too. We can't use the index, since it is changed after xSort, that's where 'ori_index' come into the picture.