Fix for #2813
Checking if the value is null before plotting the line region (dashed line). If the value is null then there is no need to add the point in the SVG path. Without this check chart plots a solid line for data points with null value. Please refer to the description in the issue #2813.